Split touch and menu bar into smaller components
