+++ src/components/electron/TouchBarPopover.js
@@ -3,7 +3,7 @@ import { forwardRef, useImperativeHandle, useLayoutEffect, useRef } from 'react'
 
 import { useElementRefValue } from '../../hooks/useElementRefValue';
 
-export const Popover = forwardRef(function Popover({
+export const TouchBarPopover = forwardRef(function TouchBarPopover({
 	children: itemsElement,
 	label,
 	icon,
