+++ src/components/electron/TouchBarSpacer.js
@@ -1,7 +1,7 @@
 import { remote } from 'electron';
 import { forwardRef, useImperativeHandle } from 'react';
 
-export const Spacer = forwardRef(function Spacer({ size }, ref) {
+export const TouchBarSpacer = forwardRef(function TouchBarSpacer({ size }, ref) {
 	useImperativeHandle(ref, () => new remote.TouchBar.TouchBarSpacer({ size }), [size]);
 
 	return null;
